Added more smart text options
Previously, it was just MAC addresses and IP addresses. I've added NetDB node names so that they are also recognized. Instead of the text expanding to more text when it is hovered upon, it now shows up as a mini popup similar to a context menu, which allows more options, looks better, and prevents a problem where text would wrap around to a new line and make it hard to click on.